+++ udata_hydra/analysis/helpers.py
@@ -41,12 +41,12 @@ async def notify_udata(resource: Record, check: dict) -> None:
         "resource_id": check["resource_id"],
         "dataset_id": resource["dataset_id"],
         "document": {
-            "analysis:parsing:error": check["parsing_error"],
+            "analysis:parsing:error": check.get("parsing_error"),
             "analysis:parsing:started_at": check["parsing_started_at"].isoformat()
-            if check["parsing_started_at"]
+            if check.get("parsing_started_at")
             else None,
             "analysis:parsing:finished_at": check["parsing_finished_at"].isoformat()
-            if check["parsing_finished_at"]
+            if check.get("parsing_finished_at")
             else None,
         },
     }